remove fixtures from code coverage report
This is a fix for issue #157.  This changes the code coverage report to
include only the resources that are declared in the module being tested
and the site manifest(s).  Previously, all resources in the catalog
generated for a particular test run were added to the coverage list
which meant that any test fixtures were also added to the coverage
report and were reported as untouched.