Adding stuff for puppet 4
Puppet 4 uses environmentpath instead of modulepath and manifest
settings. We can use a static environment based on environmentpath to
access the spec/fixtures/{manifests,modules} directories for compiling a
catalog.

The Puppet[:code] setting was always set to just "\n" for nodes on
Puppet 4. This disable reading the environment manifest .pp files.

Add a test for loading functions from fixtures.